tests: fix dtype for T4 GPU compatibility in cache accounting tests
Changed bfloat16 to float16 (half) since T4 has compute capability
7.5 and bfloat16 requires 8.0+.

+++ tests/entrypoints/openai/chat_completion/test_cache_accounting.py
@@ -10,7 +10,7 @@ from tests.utils import RemoteOpenAIServer
 def chat_server_with_prefix_caching(request):
     args = [
         "--dtype",
-        "bfloat16",
+        "half",
         "--max-model-len",
         "512",
         "--enforce-eager",
